+++ src/Controller/Management/PaymentCrudController.php
@@ -6,11 +6,15 @@ use c975L\ShopBundle\Entity\Payment;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Filters;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\EmailField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\BooleanField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IntegerField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\DateTimeField;
+use EasyCorp\Bundle\EasyAdminBundle\Filter\BooleanFilter;
+use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
+use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 
 #[IsGranted('ROLE_ADMIN')]
@@ -24,6 +28,9 @@ class PaymentCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         return [
+            AssociationField::new('basket')
+                ->setLabel('label.basket')
+                ->setFormTypeOption('disabled','disabled'),
             BooleanField::new('isFinished')
                 ->setLabel('label.is_finished'),
             IntegerField::new('amount')
@@ -38,7 +45,7 @@ class PaymentCrudController extends AbstractCrudController
                 ->setLabel('label.stripe_token_type'),
             EmailField::new('stripe_email')
                 ->setLabel('label.stripe_email'),
-                DateTimeField::new('creation')
+            DateTimeField::new('creation')
                 ->setLabel('label.creation')
                 ->hideOnIndex()
                 ->setFormTypeOption('disabled', 'disabled')
@@ -54,8 +61,9 @@ class PaymentCrudController extends AbstractCrudController
     public function configureActions(Actions $actions): Actions
     {
         return $actions
-            ->disable(Action::NEW, Action::EDIT, Action::DELETE)
+            ->disable(Action::NEW, Action::EDIT)
             ->add(Crud::PAGE_INDEX, Action::DETAIL)
+            ->setPermission(Action::DELETE, 'ROLE_ADMIN')
             ->setPermission(Action::DETAIL, 'ROLE_ADMIN')
         ;
     }
@@ -65,6 +73,13 @@ class PaymentCrudController extends AbstractCrudController
         return $crud
             ->showEntityActionsInlined()
             ->setEntityPermission('ROLE_ADMIN')
+        ;
+    }
+
+    public function configureFilters(Filters $filters): Filters
+    {
+        return $filters
+            ->add(BooleanFilter::new('isFinished'))
         ;
     }
 }