+++ src/Entity/Payment.php
@@ -56,6 +56,11 @@ class Payment
     #[ORM\ManyToOne(inversedBy: 'payments')]
     private ?User $user = null;
 
+    public function __toString()
+    {
+        return $this->id;
+    }
+
     public function getId(): int
     {
         return $this->id;