+++ src/Service/ProductService.php
@@ -20,11 +20,17 @@ class ProductService implements ProductServiceInterface
         return $this->productRepository->findAll();
     }
 
+    // Finds all
+    public function findAllSorted()
+    {
+        return $this->productRepository->findAllSorted();
+    }
+
     // Gets the products paginated
     public function findAllPaginated($query)
     {
         return $this->paginator->paginate(
-            $this->findAll(),
+            $this->findAllSorted(),
             $query->getInt('p', 1),
             9
         );